refactor: NavBar를 생성함
